refactor: make getMarkerType reusable
